+++ dislikes.php
@@ -3,17 +3,18 @@
 session_start();
 
 require_once 'src/models/PostsModel.php';
+require_once 'src/models/Likes_DislikesModel.php';
 require_once 'src/services/DatabaseConnectionService.php';
-require_once 'src/models/DislikesModel.php';
 
 $db = DatabaseConnectionService::connect();
 $PostsModel = new PostsModel($db);
-$DislikesModel = new DislikesModel($db);
+$LikesModel = new LikesModel($db);
 
 $id = $_GET['id'];
 
 if (isset($_SESSION['username'])){
-    $DislikesModel->trackDislikes($id, $_SESSION['username_id']);
+    $haveLiked = $LikesModel->trackActivity($id, $_SESSION['username_id']);
+    $LikesModel->addDislike($id, $_SESSION['username_id'], $haveLiked);
     header("Location: singlePost.php?id=$id");
 } else {
     header("Location: login.php");
