+++ likes.php
@@ -3,8 +3,9 @@
 session_start();
 
 require_once 'src/models/PostsModel.php';
+require_once 'src/models/Likes_DislikesModel.php';
+
 require_once 'src/services/DatabaseConnectionService.php';
-require_once 'src/models/LikesModel.php';
 
 $db = DatabaseConnectionService::connect();
 $PostsModel = new PostsModel($db);
@@ -13,7 +14,8 @@ $LikesModel = new LikesModel($db);
 $id = $_GET['id'];
 
 if (isset($_SESSION['username'])){
-    $LikesModel->trackLikes($id, $_SESSION['username_id']);
+    $haveLiked = $LikesModel->trackActivity($id, $_SESSION['username_id']);
+    $LikesModel->addLike($id, $_SESSION['username_id'], $haveLiked);
     header("Location: singlePost.php?id=$id");
 } else {
     header("Location: login.php");
